rename: 학기 정보 관리 관련 컴포넌트 Recruitment 폴더로 이동
