+++ src/pages/RecruitmentPage.tsx
@@ -1,5 +1,5 @@
-import RecruitmentHeader from "@/components/common/Header/RecruitmentHeader";
-import RecruitmentInfoTable from "@/components/common/Table/RecruitmentInfoTable";
+import RecruitmentHeader from "@/components/Recruitment/RecruitmentHeader";
+import RecruitmentInfoTable from "@/components/Recruitment/RecruitmentInfoTable";
 import Title from "@/components/common/Title";
 import RecruitmentContextProvider from "@/contexts/RecruitmentContext";
